add StorageBuilder to mark jars as serializable
Users can now create a StorageBuilder
and invoke `serializable` with each salsa
jar that will need to be serialized.

This allows us to ensure the TypeIds are
populated in the same order and present when
deserializing.

We will later have to add some code to check
if the definition of the type has changed
since serialization occurred since right now
we don't detect that.